+++ src/minishell1.c
@@ -14,78 +14,6 @@ static void	ft_signal_handler(int signal)
 	rl_redisplay();
 }
 
-/**
- * * Replace a set of characters on a string starting at some index
- * @param str		string to modify
- * @param charset	set of characters to replace
- * @param index		starting point of replacing
-*/
-char	*ft_replace(char *str, char const *oldset, char const *newset)
-{
-	char	*out;
-	char 	*tmp;
-	size_t	len;
-	char	*strnstr;
-
-	len = ft_strlen(str);
-	strnstr = ft_strnstr(str, oldset, len);
-	out = ft_substr(str, 0, strnstr - str);
-	tmp = ft_strjoin(out, newset);
-	free(out);
-	out = ft_strjoin(tmp, strnstr + ft_strlen(oldset));
-	free(tmp);
-	return (out);
-}
-
-static char *get_valid_ft_strchr_$(const char *str)
-{
-	char	*strchr_$;
-
-	strchr_$ = (char *)str;
-	while (1)
-	{
-		strchr_$ = ft_strchr(strchr_$, '$');
-		if (!strchr_$)
-			return (NULL);
-		if (strchr_$ == str)
-			return (strchr_$);
-		if (strchr_$[-1] != '\\')
-			return (strchr_$);
-	}
-}
-
-static void	ft_expanse_var(char **argv, t_str **env_list)
-{
-	int		i;
-	char	*strchr_$;
-	char 	*tmp;
-	size_t	len;
-	t_str 	*aux;
-
-	i = 0;
-	strchr_$ = NULL;
-	while (argv[i] != 0)
-	{
-		strchr_$ = get_valid_ft_strchr_$(argv[i]);
-		// strchr_$ = ft_strchr(argv[i], '$');
-		if (!strchr_$)
-			i++;
-		else
-		{
-			len = 1;
-			while (ft_isalnum(strchr_$[len]))
-				len++;
-			tmp = ft_substr(strchr_$, 0, len);
-			aux = ft_lst_str_get_str(env_list, &tmp[1]);
-			if (aux != NULL)
-				argv[i] = ft_replace(argv[i], tmp, ft_strchr(aux->str, '=') + 1);
-			else
-				argv[i] = ft_replace(argv[i], tmp, "");
-			free(tmp);
-		}
-	}
-}
-
 int	main(int argc, char **argv, char **env)
 {
 	char	*str_got;
@@ -115,7 +43,6 @@ int	main(int argc, char **argv, char **env)
 		else
 		{
 			param_list = modifsplit(str_got);
-			ft_expanse_var(param_list, &env_list);
 			interpreter(param_list, &env_list);
 		}
 		free(str_got);